Almost kindof fix Guess The Flag post
What in the heck!? This source parser is no good. For some reason, my
last python section will not tolerate underscores if the language of the
block is marked as python. Plain "_" causes italics to be rendered and
the "_" character is omitted. Escaping backslashes ("\_") results in
broken render output as well (even more broken!)

So, I have to mark the language in the last python block to "text", but
even then, underscores are interpreted as italics. So this last block,
for some stupid reason, requires backslash escapes on the underscores.

Don't try to escape backslashes in any of the other blocks, though! or
else the backslashes are displayed for some crazy reason.

Dumb dumb dumb